prefetcher: track overall sync status better
We always need to pass the same request into the overall sync status
checking functions, and that request has to already have its action
set to be syncing.  We also need to make sure we count the block's own
after we finish it processing it for prefetching.

So update the action first, and count the fetched bytes in a defer.